requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pyjwt==2.10.0
bcrypt==4.2.1
mysql-connector-python==9.3.0
google-genai==1.16.0
pydantic==2.10.3
python-multipart==0.0.12
